fix getLookedAtApp -ve distance bug
When selecting an application window by clicking with the mouse or pressing r, the window with the smallest distance to the camera is selected. If a window is behind the camera, this results in a negative distance and so that window is (erroniously) selected.

This commit adds a check that the distance is greater than zero to fix this.